+++ docs/guides/changelog.md
@@ -34,6 +34,7 @@ description: |-
 - Add `snooping_information_option_format_remote_id_hostname` attribute to `iosxe_dhcp` resource and data source
 - Add `iosxe_dot1x` resource and data source
 - Add `arp_timeout`, `spanning_tree_link_type` and `spanning_tree_portfast_trunk` attribute to `iosxe_interface_ethernet` resource and data source
+- Add `message_digest_keys` attribute to `iosxe_interface_ospf` resource and data source
 
 ## 0.3.3